added fibo calculation with formula

+++ Recursion/src/Fibonacii.java
@@ -2,12 +2,12 @@ public class Fibonacii {
     public static void main(String[] args){
 //        int answer=fibo(50);
 //        System.out.println(answer); this will take very much time and will not give answer
-        int answer=fiboFormula(50);
+        long answer=fiboFormula(50);
         System.out.println(answer);
 
     }
-    static int fiboFormula(int n){
-        return (int)(Math.pow((1+Math.sqrt(5))/2,n));
+    static long fiboFormula(int n){
+        return (long) (Math.pow((1+Math.sqrt(5))/2,n));
     }
     static int fibo(int n){
         //base condition
